Add change host method in Co-op game
