+++ config.json
@@ -1,0 +1,30 @@
+{
+  "uesave_path": "",
+  "save_path": "",
+  "player_list": [
+    {
+      "GUID": "",
+      "name": ""
+    },
+    {
+      "GUID": "",
+      "name": ""
+    },
+    {
+      "GUID": "",
+      "name": ""
+    },
+    {
+      "GUID": "",
+      "name": ""
+    },
+    {
+      "GUID": "",
+      "name": ""
+    },
+    {
+      "GUID": "",
+      "name": ""
+    }
+  ]
+}
